small changes in RARE timing ( grad mom style not k-space style)
small improvements in pulseqconverter.m

RARE seems now to properly export to pulseq. As directly gradient moments were put into the AREA of the gradient objects all moments should be correct.
rise time of the gradients is automatically adjusted.

last step that is missing: correct units of the gradmoms in k-space units